Replace trait with abstract class.
